Renames metadata parsing utils for clarity
Renames the `parse-data` utility module to `parse-metadata-import`
to better reflect its purpose: specifically parsing metadata for import.
This improves code clarity and maintainability.
